+++ week3-recommend-menu/src/components/common/SelectMenuBtn.jsx
@@ -1,7 +1,6 @@
 import styled, { css } from "styled-components";
 
 const SelectMenuBtn = ({ isBig = false, innerTxt, isSelected, onClick }) => {
-  console.log(isSelected);
   return (
     <>
       {isBig ? (
